Renaming and merging
